requirements.txt
flask==3.1.0
google-generativeai==0.8.4
pypdf==5.4.0